Move Laravel classes into their own namespace.
